Revert "refactor metal backend to split storage and device files for readability"
This reverts commit bb594bc1a6fb13a3bc2302f84cc4687ab7d0a74c.
